Use std::stoll for parsing 64-bit integer immediates

+++ parser/parser.cpp
@@ -86,7 +86,7 @@ void parser_t::parse_mem_instruction() {
 
     // we should be imm
     EXPECT(TOKEN_TYPE::IMMEDIATE);
-    offset_t offset = std::stoi(_current_token->word);
+    offset_t offset = std::stoll(_current_token->word);
 
     advance();
     // we should be left paranthesis
@@ -141,9 +141,9 @@ void parser_t::parse_alu_instruction() {
         EXPECT(TOKEN_TYPE::IMMEDIATE);
         is_imm = true;
         if(_current_token->word[0] == '0' && _current_token->word[1] == 'x')
-            src2_val = std::stoi(_current_token->word,nullptr,16);
+            src2_val = std::stoll(_current_token->word,nullptr,16);
         else
-            src2_val = std::stoi(_current_token->word);
+            src2_val = std::stoll(_current_token->word);
     } else if(lookup_t::alur_type(tmp.word) != alu_instruction_t::ALU_INSTRUCTION_TYPE::UNKNOWN) {
         type = lookup_t::alur_type(tmp.word);
         EXPECT(TOKEN_TYPE::REGISTER);
@@ -243,9 +243,9 @@ void parser_t::parse_jump_instruction() {
     EXPECT(TOKEN_TYPE::IMMEDIATE);
     int64_t imm;
     if(_current_token->word[0] == '0' && _current_token->word[1] == 'x')
-        imm = std::stoi(_current_token->word,nullptr,16);
+        imm = std::stoll(_current_token->word,nullptr,16);
     else
-        imm= std::stoi(_current_token->word);
+        imm= std::stoll(_current_token->word);
     // here we dont care about the label_id field
     _program.emplace_back(
         std::make_unique<jump_instruction_t>(
@@ -270,9 +270,9 @@ void parser_t::parse_upperimm_instruction() {
     int64_t imm;
     if(_current_token->word[0] == '0' && _current_token->word[1] == 'x')
 
-        imm = std::stoi(_current_token->word,nullptr,16);
+        imm = std::stoll(_current_token->word,nullptr,16);
     else
-        imm= std::stoi(_current_token->word);
+        imm= std::stoll(_current_token->word);
     if(tmp.word == "lui")
         _program.emplace_back(std::make_unique<load_upper_imm_instruction_t>(
             dest_reg,
